More corrections to fix tests and resolve circular dependency problems; new selector for FrontsList
