+++ client-v2/src/actions/Fronts.ts
@@ -1,24 +1,12 @@
 import { Dispatch, ThunkResult } from 'types/Store';
 import { Action } from 'types/Action';
-import { batchActions } from 'redux-batched-actions';
 import {
   fetchFrontsConfig,
-  fetchLastPressed as fetchLastPressedApi,
-  publishCollection as publishCollectionApi,
-  getCollection as getCollectionApi
+  fetchLastPressed as fetchLastPressedApi
 } from 'services/faciaApi';
 import { actions as frontsConfigActions } from 'bundles/frontsConfigBundle';
-import { recordUnpublishedChanges } from 'actions/UnpublishedChanges';
-import { isFrontStale } from 'util/frontsUtils';
-import { getCollections, getArticlesForCollections } from 'actions/Collections';
 import { VisibleArticlesResponse } from 'types/FaciaApi';
-import { visibleArticlesSelector, getFront } from 'selectors/frontsSelectors';
-import { Stages, CollectionItemSets } from 'shared/types/Collection';
-import { frontStages, noOfOpenCollectionsOnFirstLoad } from 'constants/fronts';
-import { State } from 'types/State';
-import { editorOpenCollections } from 'bundles/frontsUIBundle';
-import { events } from 'services/GA';
-import { collectionParamsSelector } from 'selectors/collectionSelectors';
+import { Stages } from 'shared/types/Collection';
 
 function fetchLastPressedSuccess(frontId: string, datePressed: string): Action {
   return {
@@ -75,98 +63,12 @@ function fetchLastPressed(frontId: string): ThunkResult<void> {
       });
 }
 
-function publishCollection(
-  collectionId: string,
-  frontId: string
-): ThunkResult<Promise<void>> {
-  events.collectionPublished(frontId, collectionId);
-
-  return (dispatch: Dispatch, getState: () => State) => {
-    const draftVisibleArticles = visibleArticlesSelector(getState(), {
-      collectionId,
-      stage: frontStages.draft
-    });
-
-    return publishCollectionApi(collectionId)
-      .then(() => {
-        dispatch(
-          batchActions([
-            publishCollectionSuccess(collectionId),
-            recordUnpublishedChanges(collectionId, false),
-            recordVisibleArticles(
-              collectionId,
-              draftVisibleArticles,
-              frontStages.live
-            )
-          ])
-        );
-
-        dispatch(getCollections([collectionId]));
-
-        return new Promise(resolve => setTimeout(resolve, 10000))
-          .then(() => {
-            const [params] = collectionParamsSelector(getState(), [
-              collectionId
-            ]);
-            return Promise.all([
-              getCollectionApi(params),
-              fetchLastPressedApi(frontId)
-            ]);
-          })
-          .then(([collectionResponse, lastPressed]) => {
-            const lastPressedInMilliseconds = new Date(lastPressed).getTime();
-            dispatch(
-              batchActions([
-                recordStaleFronts(
-                  frontId,
-                  isFrontStale(
-                    collectionResponse.collection.lastUpdated,
-                    lastPressedInMilliseconds
-                  )
-                ),
-                fetchLastPressedSuccess(frontId, lastPressed)
-              ])
-            );
-          });
-      })
-      .catch(() => {
-        // @todo: implement once error handling is done
-      });
-  };
-}
-
-/**
- * Initialise a front --
- * - Fetch all of its collections from the server
- * - Mark as open number of collections indicated by the constant, and fetch their articles
- */
-function initialiseFront(
-  frontId: string,
-  browsingStage: CollectionItemSets
-): ThunkResult<Promise<void>> {
-  return async (dispatch: Dispatch, getState: () => State) => {
-    const front = getFront(getState(), frontId);
-    if (!front) {
-      return;
-    }
-    const collectionsWithArticlesToLoad = front.collections.slice(
-      0,
-      noOfOpenCollectionsOnFirstLoad
-    );
-    dispatch(editorOpenCollections(collectionsWithArticlesToLoad));
-    await dispatch(getCollections(front.collections));
-    await dispatch(
-      getArticlesForCollections(collectionsWithArticlesToLoad, browsingStage)
-    );
-  };
-}
-
 export {
   fetchLastPressed,
   fetchLastPressedSuccess,
-  publishCollection,
   recordVisibleArticles,
-  initialiseFront
+  publishCollectionSuccess,
+  recordStaleFronts
 };
 
 export default function getFrontsConfig(): ThunkResult<
