+++ client-v2/src/bundles/frontsUIBundle.ts
@@ -22,13 +22,12 @@ import {
 } from 'types/Action';
 import { State as GlobalState } from 'types/State';
 import { events } from 'services/GA';
-import { getFronts } from 'selectors/frontsSelectors';
+import { getFronts, getFrontsWithPriority } from 'selectors/frontsSelectors';
 import { createSelector } from 'reselect';
 import {
   REMOVE_GROUP_ARTICLE_FRAGMENT,
   REMOVE_SUPPORTING_ARTICLE_FRAGMENT
 } from 'shared/actions/ArticleFragments';
-import { REMOVE_CLIPBOARD_ARTICLE_FRAGMENT } from 'actions/Clipboard';
 
 export const EDITOR_OPEN_CURRENT_FRONTS_MENU =
   'EDITOR_OPEN_CURRENT_FRONTS_MENU';
@@ -231,6 +230,21 @@ const createSelectEditorFrontsByPriority = () =>
     }
   );
 
+const createSelectFrontIdAndOpenStateByPriority = () => {
+  const selectEditorFrontsByPriority = createSelectEditorFrontsByPriority();
+  return createSelector(
+    getFrontsWithPriority,
+    (state, priority: string) =>
+      selectEditorFrontsByPriority(state, { priority }),
+    (frontsForPriority, openFronts) => {
+      return frontsForPriority.map(({ id }) => ({
+        id,
+        isOpen: !!openFronts.find(_ => _.id === id)
+      }));
+    }
+  );
+};
+
 const selectEditorFrontIds = (state: GlobalState) =>
   state.editor.frontIdsByPriority;
 
@@ -369,7 +383,7 @@ const reducer = (state: State = defaultState, action: Action): State => {
     }
     case REMOVE_SUPPORTING_ARTICLE_FRAGMENT:
     case REMOVE_GROUP_ARTICLE_FRAGMENT:
-    case REMOVE_CLIPBOARD_ARTICLE_FRAGMENT: {
+    case 'REMOVE_CLIPBOARD_ARTICLE_FRAGMENT': {
       const articleFragmentId = action.payload.articleFragmentId;
       const selectedFrontId = Object.keys(state.selectedArticleFragments).find(
         frontId => {
@@ -443,6 +457,7 @@ export {
   selectIsCurrentFrontsMenuOpen,
   selectEditorFrontIds,
   createSelectEditorFrontsByPriority,
+  createSelectFrontIdAndOpenStateByPriority,
   selectEditorFrontIdsByPriority,
   selectEditorArticleFragment,
   selectIsCollectionOpen,
