+++ client-v2/src/components/FrontsEdit/Front.tsx
@@ -29,11 +29,11 @@ import { getFront } from 'selectors/frontsSelectors';
 import { FrontConfig } from 'types/FaciaApi';
 import { visibleFrontArticlesSelector } from 'selectors/frontsSelectors';
 import { VisibleArticlesResponse } from 'types/FaciaApi';
-import { initialiseFront } from 'actions/Fronts';
 import { events } from 'services/GA';
 import FrontDetailView from './FrontDetailView';
 import CollectionItem from './CollectionComponents/CollectionItem';
 import { ValidationResponse } from 'shared/util/validateImageSrc';
+import { initialiseCollectionsForFront } from 'actions/Collections';
 
 const FrontContainer = styled('div')`
   display: flex;
@@ -262,7 +262,7 @@ const mapDispatchToProps = (
   return {
     dispatch,
     initialiseFront: () =>
-      dispatch(initialiseFront(props.id, props.browsingStage)),
+      dispatch(initialiseCollectionsForFront(props.id, props.browsingStage)),
     selectArticleFragment: (
       frontId: string,
       articleFragmentId: string,
